Refactored PaymentsResponse and PaymentsDetailsResponse for less duplicate code.
